Extract layout utils
